More flexible deployment outside of a common App Bundle on MacOS
Generally, CEF/JCEF applications need to be deployed with the helper app
and the framework inside an application bundle on MacOS. This strict
directory structure is necessary because the dynamic CEF loading
mechanism relies on it, and because the Mach IPC registration mechanism
for IPC channels between the helper process and the master process uses
the outer bundle ID for identification.
This patch implements fixes/workarounds for both of these problems,
effectively relaxing the requirement and allowing effective wrapping of
JCEF and CEF on MacOS within a JAR that extracts the files at runtime to
a temporary location.

Cherry-pick of 8eadc1495a2e58fde7e7fefb44870fec58d7f765 from branch
S1artie/3578.